requirements.txt
annotated-types==0.6.0
anyio==4.3.0
apify_client==1.6.4
apify_shared==1.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
FLAML==2.1.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.47
numpy==1.26.4
openai==1.19.0
packaging==24.0
pyautogen==0.2.24
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.1
tavily-python==0.3.3
termcolor==2.4.0
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
